refactor: move wait for cluster function to conftest
